refactor: merge preferences.py into agents_md.py
Preference dimensions, instructions, and render_preferences() all
describe AGENTS.md content — one module owns the full picture.
Delete wos/preferences.py and update all import sites.
